Move IDocumentPositionInfoStrategy and friends to Workspaces layer
